+++ FoodMinder/src/main/java/com/foodminder/FoodMinder/services/RefeicaoService.java
@@ -1,35 +1,42 @@
 package com.foodminder.FoodMinder.services;
 
-import com.foodminder.FoodMinder.domain.planejamento.PlanejamentoRepository;
 import com.foodminder.FoodMinder.domain.refeicao.Refeicao;
 import com.foodminder.FoodMinder.domain.refeicao.RefeicaoRepository;
 import com.foodminder.FoodMinder.domain.refeicao.RequestRefeicao;
-import com.foodminder.FoodMinder.exceptions.RecursoNaoEncontrado;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
-import java.util.List;
+import java.util.Optional;
 
 @Service
 public class RefeicaoService {
     @Autowired
-    private PlanejamentoRepository planejamentoRepository;
-    @Autowired
     private RefeicaoRepository refeicaoRepository;
-    public List<Refeicao> obterTodasRefeicoes() {
-        return refeicaoRepository.findAll();
+    public ResponseEntity obterTodasRefeicoes() {
+        return ResponseEntity.ok(refeicaoRepository.findAll());
     }
-    public Refeicao obterRefeicaoPorId(Integer id) {
-        return refeicaoRepository.findById(id)
-                .orElseThrow(() -> new RecursoNaoEncontrado("Planejamento não encontrado"));
+    public ResponseEntity obterRefeicaoPorId(Integer id) {
+        Optional<Refeicao> refeicaoOptional = refeicaoRepository.findById(id);
+        if (refeicaoOptional.isPresent()) {
+            Refeicao refeicao = refeicaoOptional.get();
+            return ResponseEntity.ok(refeicao);
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
-    public Refeicao registrarRefeicao(RequestRefeicao data) {
-        Refeicao newRefeicao = new Refeicao(data);
-        return(refeicaoRepository.save(newRefeicao));
+    public ResponseEntity registrarRefeicao(RequestRefeicao requestRefeicao) {
+        Refeicao refeicao = new Refeicao(requestRefeicao);
+        refeicaoRepository.save(refeicao);
+        return ResponseEntity.status(HttpStatus.CREATED).body(refeicao);
     }
-    public Refeicao deletarRefeicaoPorId(Integer id) {
-        Refeicao refeicao = refeicaoRepository.findById(id)
-                .orElseThrow(() -> new RecursoNaoEncontrado("Planejamento não encontrado"));
-        refeicaoRepository.deleteById(id);
-        return refeicao;
+    public ResponseEntity deletarRefeicaoPorId(Integer id) {
+        Optional<Refeicao> refeicaoOptional = refeicaoRepository.findById(id);
+        if (refeicaoOptional.isPresent()) {
+            refeicaoRepository.deleteById(id);
+            return ResponseEntity.noContent().build();
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
